Be more specific with background window hints
Explicitly set the hints that should be implied by the desktop window
type. This will probably be a noop on most WMs, but it doesn't hurt to
set them anyway.

Also, attempt to set the desktop window type via Qt attributes. This
doesn't seem to set the atom for some reason (not with Qt 5.6 or 5.7 at
least), so the libxcb code is still needed, but again, it doesn't hurt
to set it anyway.